requirements.txt
flask
requests
pyjwt
mysql-connector-python
pandas